requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
defusedxml==0.7.1
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
httpie==2.6.0
idna==3.3
importlib-metadata==4.10.0
Markdown==3.3.6
Pygments==2.11.2
PySocks==1.7.1
python-frontmatter==1.0.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
sqlparse==0.4.2
urllib3==1.26.8
zipp==3.7.0
